+++ controller/konnect/watch_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/kong/gateway-operator/controller/konnect/constraints"
 	"github.com/kong/gateway-operator/modules/manager/scheme"
 
+	commonv1alpha1 "github.com/kong/kubernetes-configuration/api/common/v1alpha1"
 	configurationv1 "github.com/kong/kubernetes-configuration/api/configuration/v1"
 	configurationv1alpha1 "github.com/kong/kubernetes-configuration/api/configuration/v1alpha1"
 	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
@@ -135,7 +136,7 @@ func TestEnqueueObjectForKonnectGatewayControlPlane(t *testing.T) {
 							Namespace: "default",
 						},
 						Spec: configurationv1.KongConsumerSpec{
-							ControlPlaneRef: &configurationv1alpha1.ControlPlaneRef{
+							ControlPlaneRef: &commonv1alpha1.ControlPlaneRef{
 								Type: configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef,
 								KonnectNamespacedRef: &configurationv1alpha1.KonnectNamespacedRef{
 									Name: cp.Name,
@@ -170,7 +171,7 @@ func TestEnqueueObjectForKonnectGatewayControlPlane(t *testing.T) {
 							Namespace: "default",
 						},
 						Spec: configurationv1.KongConsumerSpec{
-							ControlPlaneRef: &configurationv1alpha1.ControlPlaneRef{
+							ControlPlaneRef: &commonv1alpha1.ControlPlaneRef{
 								Type: configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef,
 								KonnectNamespacedRef: &configurationv1alpha1.KonnectNamespacedRef{
 									Name: "different-cp",