+++ test/envtest/condition_asserts.go
@@ -5,7 +5,7 @@ import (
 
 	k8sutils "github.com/kong/gateway-operator/pkg/utils/kubernetes"
 
-	configurationv1alpha1 "github.com/kong/kubernetes-configuration/api/configuration/v1alpha1"
+	commonv1alpha1 "github.com/kong/kubernetes-configuration/api/common/v1alpha1"
 	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 )
 
@@ -13,14 +13,14 @@ func conditionsAreSetWhenReferencedControlPlaneIsMissing[
 	T interface {
 		client.Object
 		k8sutils.ConditionsAware
-		GetControlPlaneRef() *configurationv1alpha1.ControlPlaneRef
+		GetControlPlaneRef() *commonv1alpha1.ControlPlaneRef
 	},
 ](objToMatch T) func(obj T) bool {
 	return func(obj T) bool {
 		if obj.GetName() != objToMatch.GetName() {
 			return false
 		}
-		if obj.GetControlPlaneRef().Type != configurationv1alpha1.ControlPlaneRefKonnectID {
+		if obj.GetControlPlaneRef().Type != commonv1alpha1.ControlPlaneRefKonnectID {
 			return false
 		}
 		condCpRef, okCPRef := k8sutils.GetCondition(konnectv1alpha1.ControlPlaneRefValidConditionType, obj)
@@ -37,7 +37,7 @@ func conditionProgrammedIsSetToTrue[
 	T interface {
 		client.Object
 		k8sutils.ConditionsAware
-		GetControlPlaneRef() *configurationv1alpha1.ControlPlaneRef
+		GetControlPlaneRef() *commonv1alpha1.ControlPlaneRef
 		GetKonnectID() string
 	},
 ](objToMatch T, id string) func(T) bool {
@@ -45,7 +45,7 @@ func conditionProgrammedIsSetToTrue[
 		if obj.GetName() != objToMatch.GetName() {
 			return false
 		}
-		if obj.GetControlPlaneRef().Type != configurationv1alpha1.ControlPlaneRefKonnectID {
+		if obj.GetControlPlaneRef().Type != commonv1alpha1.ControlPlaneRefKonnectID {
 			return false
 		}
 		return obj.GetKonnectID() == id && k8sutils.IsProgrammed(obj)